+++ retdec/decompiler.py
@@ -27,6 +27,9 @@ class Decompiler(Service):
         :type mode: str
         :param target_language: Target high-level language.
         :type target_language: str
+        :param graph_format: Format of the generated call and control-flow
+            graphs.
+        :type graph_format: str
         :param decomp_var_names: Naming style for variables.
         :type decomp_var_names: str
         :param decomp_optimizations: Level of optimizations performed by the
@@ -68,6 +71,11 @@ class Decompiler(Service):
         :param sel_decomp_decoding: What instructions should be decoded when
             either `sel_decomp_funcs` or `sel_decomp_ranges` is given?
         :type sel_decomp_decoding: str
+        :param generate_cg: Should a call graph be generated?
+        :type generate_cg: bool
+        :param generate_cfgs: Should control-flow graphs for all functions be
+            generated?
+        :type generate_cfgs: bool
         :param generate_archive: Should an archive containing all outputs from
             the decompilation be generated?
         :type generate_archive: bool
@@ -104,6 +112,7 @@ class Decompiler(Service):
             'mode': self._get_mode_param(files['input'], kwargs)
         }
         self._add_param_when_given('target_language', params, kwargs)
+        self._add_param_when_given('graph_format', params, kwargs)
         self._add_param_when_given('decomp_var_names', params, kwargs)
         self._add_param_when_given('decomp_optimizations', params, kwargs)
         self._add_param_when_given('decomp_unreach_funcs', params, kwargs)
@@ -118,6 +127,8 @@ class Decompiler(Service):
         self._add_sel_decomp_ranges_param_when_given(params, kwargs)
         self._add_param_when_given('sel_decomp_decoding', params, kwargs)
         self._add_param_when_given('generate_archive', params, kwargs)
+        self._add_param_when_given('generate_cg', params, kwargs)
+        self._add_param_when_given('generate_cfgs', params, kwargs)
         response = conn.send_post_request(files=files, params=params)
         return response['id']