+++ tests/decompiler_tests.py
@@ -133,6 +133,15 @@ class DecompilerStartDecompilationTests(BaseServiceTests):
             params=AnyParamsWith(target_language='py')
         )
 
+    def test_graph_format_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            graph_format='svg'
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(graph_format='svg')
+        )
+
     def test_decomp_var_names_is_set_to_correct_value_when_given(self):
         self.start_decompilation_with_any_input_file(
             decomp_var_names='simple'
@@ -299,6 +308,24 @@ class DecompilerStartDecompilationTests(BaseServiceTests):
 
         self.assert_post_request_was_sent_with(
             params=AnyParamsWith(sel_decomp_decoding='only')
+        )
+
+    def test_generate_cg_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            generate_cg=True
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(generate_cg=True)
+        )
+
+    def test_generate_cfgs_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            generate_cfgs=True
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(generate_cfgs=True)
         )
 
     def test_generate_archive_is_set_to_correct_value_when_given(self):