+++ tests/exceptions_tests.py
@@ -11,11 +11,14 @@ import unittest
 from retdec.exceptions import AnalysisFailedError
 from retdec.exceptions import ArchiveGenerationFailedError
 from retdec.exceptions import AuthenticationError
+from retdec.exceptions import CFGGenerationFailedError
+from retdec.exceptions import CGGenerationFailedError
 from retdec.exceptions import ConnectionError
 from retdec.exceptions import DecompilationFailedError
 from retdec.exceptions import InvalidValueError
 from retdec.exceptions import MissingAPIKeyError
 from retdec.exceptions import MissingParameterError
+from retdec.exceptions import NoSuchCFGError
 from retdec.exceptions import OutputNotRequestedError
 from retdec.exceptions import UnknownAPIError
 
@@ -97,6 +100,42 @@ class OutputNotRequestedErrorTests(unittest.TestCase):
     def test_has_correct_description(self):
         ex = OutputNotRequestedError()
         self.assertIn('not requested', str(ex))
+
+
+class CGGenerationFailedErrorTests(unittest.TestCase):
+    """Tests for :class:`retdec.exceptions.CGGenerationFailedError`."""
+
+    def test_does_not_include_reason_when_not_given(self):
+        ex = CGGenerationFailedError()
+        self.assertNotIn('reason', str(ex))
+
+    def test_includes_reason_when_given(self):
+        ex = CGGenerationFailedError('REASON')
+        self.assertIn('REASON', str(ex))
+
+
+class CFGGenerationFailedErrorTests(unittest.TestCase):
+    """Tests for :class:`retdec.exceptions.CFGGenerationFailedError`."""
+
+    def test_includes_func_name(self):
+        ex = CFGGenerationFailedError('my_func')
+        self.assertIn('my_func', str(ex))
+
+    def test_does_not_include_reason_when_not_given(self):
+        ex = CFGGenerationFailedError('my_func')
+        self.assertNotIn('reason', str(ex))
+
+    def test_includes_reason_when_given(self):
+        ex = CFGGenerationFailedError('my_func', 'REASON')
+        self.assertIn('REASON', str(ex))
+
+
+class NoSuchCFGErrorTests(unittest.TestCase):
+    """Tests for :class:`retdec.exceptions.NoSuchCFGError`."""
+
+    def test_includes_func_name(self):
+        ex = NoSuchCFGError('my_func')
+        self.assertIn('my_func', str(ex))
 
 
 class ArchiveGenerationFailedErrorTests(unittest.TestCase):