+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Changelog
 dev
 ---
 
+* Added support for generating and downloading of control-flow graphs.
+* Added support for generating and downloading of a call graph.
+* Added support for selecting the format of the generated call and control-flow
+  graphs.
 * Added support for selecting a different style for naming of variables.
 * Added support for selecting the type of optimizations to be performed by the
   decompiler.